+++ CHANGELOG.md
@@ -4,4 +4,4 @@ IMPROVEMENTS:
 
 * Added github templates for PRs, Issues etc. ([#133](https://github.com/SentientTechnologies/studio-go-runner/issue/133)).
 * Capture artifact downloading failures and insert them into the experiments output file. ([#133](https://github.com/SentientTechnologies/studio-go-runner/issue/133)).
-
+* Migrate official open source Nvidia NVML go lang bindings for GPU device management